pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "realmforge"
version = "0.1.0"
description = "Domain-agnostic energy-based world model backbone with BYOC domain overlays"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "RealmForge Team"}]
dependencies = [
  "pandas>=2.1",
  "numpy>=1.26",
  "pyarrow>=14.0",
  "scikit-learn>=1.4",
  "pydantic>=2.7",
  "pyyaml>=6.0",
  "torch>=2.2",
  "lightning>=2.2",
  "einops>=0.8",
  "fastapi>=0.111",
  "uvicorn>=0.30",
  "duckdb>=1.0",
  "umap-learn>=0.5",
  "matplotlib>=3.8"
]

[project.scripts]
realm = "app.cli:main"

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-cov>=5.0",
  "ruff>=0.6.8",
  "mypy>=1.11",
  "bandit[toml]>=1.7.9",
  "pre-commit>=3.8",
  "types-PyYAML>=6.0.12"
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "realms*"]
exclude = ["tests*", "notebooks*", "dist*", "build*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
addopts = "-q --disable-warnings --maxfail=1"

[tool.ruff]
line-length = 100
target-version = "py310"
exclude = ["notebooks"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E203"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
check_untyped_defs = true
exclude = ["notebooks/"]

[tool.bandit]
exclude_dirs = ["tests", "notebooks", ".venv"]
skips = ["B101"]
